feat: apply `sort-config` rule by default to Nuxt config files in `.config`
Nuxt supports the [config-dir proposal](https://github.com/pi0/config-dir) in the code, though it seems to be left out of the docs at the moment.

This commit only directly supports the convention for configs as defined in the proposal (i.e. `.config/<toolname>.<ext>`). For other filenames, users can use the rule in their own ESLint configs